03 feature

+++ file-storage-service/src/main/java/ru/answer_42/file_storage_service/service/impl/StorageServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.answer_42.file_storage_service.service.impl;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -12,8 +13,10 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ArrayUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
@@ -30,27 +33,40 @@ import ru.answer_42.file_storage_service.exception.storage.StorageLocationEmptyE
 import ru.answer_42.file_storage_service.exception.storage.StorageStoreFailedException;
 import ru.answer_42.file_storage_service.mapper.FileMapper;
 import ru.answer_42.file_storage_service.model.File;
+import ru.answer_42.file_storage_service.model.RemovedFiles;
 import ru.answer_42.file_storage_service.model.Status;
+import ru.answer_42.file_storage_service.repository.FileRepository;
 import ru.answer_42.file_storage_service.service.AntivirusService;
 import ru.answer_42.file_storage_service.service.FileService;
+import ru.answer_42.file_storage_service.service.RemovedFilesService;
 import ru.answer_42.file_storage_service.service.StorageService;
 
 @Service
+@Slf4j
 public class StorageServiceImpl implements StorageService {
 
   public final static long MIN_SIZE = 1L;
-  public final static long MAX_SIZE = 15 * 1024 * 1024L;
+  public final static long MAX_SIZE = 15 * 1024 * 1024;
   private final Path rootLocation;
   private final FileService fileService;
   private final AntivirusService antivirusService;
   private final FileMapper fileMapper;
+  private final RemovedFilesService removedFilesService;
+
+  @Autowired
+  @Qualifier("commonExecutor")
+  private Executor executor;
+  @Autowired
+  private FileRepository fileRepository;
 
   @Autowired
   public StorageServiceImpl(StorageProperties properties, FileService fileService,
-      AntivirusService antivirusService, FileMapper fileMapper) {
+      AntivirusService antivirusService, FileMapper fileMapper,
+      RemovedFilesService removedFilesService) {
     if (properties.getLocation().trim().isEmpty()) {
       throw new StorageLocationEmptyException("File upload location can not be Empty.");
     }
+    this.removedFilesService = removedFilesService;
     this.rootLocation = Paths.get(properties.getLocation());
     this.fileService = fileService;
     this.antivirusService = antivirusService;
@@ -60,49 +76,58 @@ public class StorageServiceImpl implements StorageService {
 
   @Override
   public FileResponseDto store(UUID userId, MultipartFile file) {
-    Executor executor = Executors.newFixedThreadPool(10);
-    try {
-      if (file.isEmpty()) {
-        throw new FileIsEmptyException("Failed to store empty file.");
-      }
-      Path destinationFile = this.rootLocation.resolve(
-              userId.toString()).resolve(Paths.get(file.getOriginalFilename()))
-          .normalize();
-      if (file.getSize() > MAX_SIZE) {
-        throw new FileSizeLimitExceededException("File is too large");
-      }
 
-      FileResponseDto fileEntity = fileService.multipartFileToFileResponseDto(userId, file,
-          destinationFile);
-      FileResponseDto fileResponseDto = fileService.save(userId,
-          fileMapper.toFileRequestDtoFromFileResponseDto(fileEntity));
-      UUID fileId = fileService.getFileIdByUserIdAndTitle(userId, fileResponseDto.getTitle());
-      CompletableFuture future = CompletableFuture.runAsync(() -> {
-        fileService.updateStatus(fileId, Status.IN_PROCESS);
-        if (!antivirusService.scan(file)) {
-          throw new RuntimeException();
-        }
-      }, executor).exceptionally(ex -> {
-        fileService.deleteById(fileId);
+    if (file.isEmpty()) {
+      throw new FileIsEmptyException("Failed to store empty file.");
+    }
+    Path destinationFile = this.rootLocation.resolve(
+            userId.toString()).normalize();
+    if (file.getSize() > MAX_SIZE) {
+      throw new FileSizeLimitExceededException("File is too large");
+    }
+
+    FileResponseDto fileEntity = fileService.multipartFileToFileResponseDto(userId, file);
+    FileResponseDto fileResponseDto = fileService.save(userId,
+        fileMapper.toFileRequestDtoFromFileResponseDto(fileEntity));
+    UUID fileId = fileService.getFileIdByUserIdAndTitle(userId, fileResponseDto.getTitle());
+    fileResponseDto = fileService.setUpDownloadUrl(fileId, destinationFile);
+    FileResponseDto finalFileResponseDto = fileResponseDto;
+    CompletableFuture.runAsync(() -> {
+      fileService.updateStatus(fileId, Status.IN_PROCESS);
+      if (!antivirusService.scan(file)) {
         throw new FileHasVirusException("File: " + file.getOriginalFilename() + " - has a virus");
-      }).thenRunAsync(() -> {
-        fileService.updateStatus(fileId, Status.READY);
-      }, CompletableFuture.delayedExecutor(1, TimeUnit.SECONDS, executor));
-      future.join();
-      fileEntity.setStatus(Status.READY);
+      }
+    }, executor).handle((result, ex) -> {
+      if(ex != null){
+        removedFilesService.save(new RemovedFiles(fileId, finalFileResponseDto.getTitle()));
+        fileService.deleteById(fileId);
+        fileService.updateStatus(fileId, Status.HAS_A_VIRUS);
+        log.error("Virus detected", ex);
+        return Status.HAS_A_VIRUS;
+      }
+      else {
+        return Status.IN_PROCESS;
+      }
+    }).thenAcceptAsync(status -> {
+      if (status == Status.HAS_A_VIRUS) {
+        return;
+      }
+      fileService.updateStatus(fileId, Status.READY);
+      finalFileResponseDto.setStatus(Status.READY);
       FileOrder fileOrder =  fileMapper.toFileMetadataOrderFromFileResponseDto(fileEntity);
       fileOrder.setUserId(userId);
       fileOrder.setId(fileId);
-      fileService.createFileOrder(fileOrder);
-      return fileEntity;
-    } catch (IOException e) {
-      throw new StorageStoreFailedException("Failed to store file.");
-    }
+      try {
+        fileService.createFileOrder(fileOrder);
+      } catch (JsonProcessingException e) {
+        throw new RuntimeException(e);
+      }
+    }, CompletableFuture.delayedExecutor(15, TimeUnit.SECONDS, executor));
+    return fileEntity;
   }
 
   @Override
   public Resource loadAll(UUID userId, List<UUID> fileNames) {
-    Executor executor = Executors.newFixedThreadPool(10);
     CompletableFuture<List<FileResponseDto>> futures = CompletableFuture.supplyAsync(
         () -> fileService.findByUserIdAndFilesId(userId, fileNames), executor);
     List<FileResponseDto> desiredFiles = futures.join();
@@ -111,16 +136,14 @@ public class StorageServiceImpl implements StorageService {
       for (FileResponseDto file : desiredFiles) {
         ZipEntry entry1 = new ZipEntry(file.getTitle());
         zout.putNextEntry(entry1);
-        // считываем содержимое файла в массив byte
-        byte[] buffer = file.getFile();
         // добавляем содержимое к архиву
-        zout.write(buffer);
+        zout.write(file.getFile());
         // закрываем текущую запись для новой записи
         zout.closeEntry();
 
       }
     } catch (Exception ex) {
-      System.out.println(ex.getMessage());
+      log.info(ex.getMessage());
     }
     ByteArrayResource byteArrayResource = new ByteArrayResource(baos.toByteArray());
     return byteArrayResource;
